requirements.txt
matplotlib==3.0.3
numpy==1.16.3
pandas==0.24.2
PyYAML==5.1
scipy==1.2.1
seaborn==0.9.0
tensorboardX==1.6
torch==1.1.0
torchvision==0.2.2.post3
